renamed reducers and actions files
