+++ src/components/reducers.test.js
@@ -4,7 +4,7 @@ import {
   CATEGORY_ADD,
   CATEGORY_UPDATE,
   CATEGORY_REMOVE
-} from './reducers';
+} from './category-reducers';
 
 describe('categories reducers', () => {
